developemnt support script for recompiling and optionally running coverage tests

+++ developmentSupportAndScripts/resetAndInitiateJestCoverageTest.ps1
@@ -1,3 +1,7 @@
+param(
+    [Boolean]$runJestTests
+)
+
 if ((uname) -eq 'Darwin') {
     Write-Output "macOS"
     Set-Alias start open
@@ -9,17 +13,20 @@ If (Test-Path $compileDirectory ) {
     Remove-Item $compileDirectory -Recurse -Force
 }
 
-### CLEAR OUT LATEST CODE COVERAGE ARTIFACTS
-$coverageDirectory = "./coverage/"
-If (Test-Path $coverageDirectory ) {
-    Remove-Item $coverageDirectory -Recurse -Force
-}
-
 ### RECOMPILE
 npm run compile
 
-### RUN CONVERAGE TESTS
-npm run jest-test
+if ( $runJestTests ) {
+    ### CLEAR OUT LATEST CODE COVERAGE ARTIFACTS
+    $coverageDirectory = "./coverage/"
+    If (Test-Path $coverageDirectory ) {
+        Remove-Item $coverageDirectory -Recurse -Force
+    }
 
-### OPEN UP COVERAGE REPORT IN BROWSER BASED ON EXPECTED ROOT DIRECTORY CREATION OF COVERAGE FOLDER
-start ./coverage/lcov-report/index.html
+    ### RUN CONVERAGE TESTS
+    npm run jest-test
+
+    ### OPEN UP COVERAGE REPORT IN BROWSER BASED ON EXPECTED ROOT DIRECTORY CREATION OF COVERAGE FOLDER
+    start ./coverage/lcov-report/index.html
+    
+}
